More lexer designs isolations
