+++ src/txxt/lexer/linebased.rs
@@ -5,7 +5,10 @@
 //! - Transforms line tokens into a hierarchical tree
 
 pub mod pipeline;
+pub mod tokens;
+pub mod transformations;
 
 pub use pipeline::{
     experimental_lex, experimental_lex_stage, PipelineError, PipelineOutput, PipelineStage,
 };
+pub use tokens::{LineToken, LineTokenTree, LineTokenType};
