+++ src/txxt/lexer/linebased/transformations.rs
@@ -1,0 +1,11 @@
+//! Line-based transformation modules
+//!
+//! This module contains transformations specific to the line-based lexer pipeline:
+//! - transform_to_line_tokens: Flat tokens → line tokens
+//! - transform_indentation_to_token_tree: Line tokens → hierarchical tree
+
+pub mod transform_indentation_to_token_tree;
+pub mod transform_to_line_tokens;
+
+pub use transform_indentation_to_token_tree::experimental_transform_indentation_to_token_tree;
+pub use transform_to_line_tokens::experimental_transform_to_line_tokens;
